fix warning error in HypCommonFunc::flock_put_contents()
動作に支障はないが、環境により Warning エラーが出る問題の修正。 (Thx nonn50)

+++ xoops_trust_path/class/hyp_common/hyp_common_func.php
@@ -2007,7 +2007,7 @@ EOD;
 				if (file_put_contents($temp, $src)) {
 					$return = @ rename($temp, $filename);
 				}
-				unlink($temp);
+				if (! $return) unlink($temp);
 			}
 			if (! $return && $handle = @ fopen($filename, $mode)) {
 				$i = 0;
